requirements.txt
aiofiles==22.1.0
allure-behave==2.12.0
allure-python-commons==2.12.0
Appium-Python-Client==2.7.1
async-generator==1.10
attrs==22.1.0
behave==1.2.6
behave-html-formatter==0.9.10
behave-testrail-reporter==0.5.1
behavex==1.6.0
certifi==2022.9.24
cffi==1.15.1
chromedriver-binary-auto==0.2.3
click==8.1.3
configobj==5.0.6
cryptography==38.0.4
csscompressor==0.9.5
decorator==5.1.1
Faker==15.3.4
fastjsonschema==2.16.2
fb-idb==1.1.7
future==0.18.2
gitdb==4.0.10
GitPython==3.1.29
grpclib==0.4.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htmlmin==0.1.12
hyperframe==6.0.1
idna==3.4
iniconfig==1.1.1
invoke==1.7.3
Jinja2==3.1.2
jsonschema==4.17.3
jupyter-core==5.1.0
MarkupSafe==2.1.1
mock==4.0.3
multidict==6.0.3
nbformat==5.7.0
nose==1.3.7
outcome==1.2.0
packaging==21.3
parse==1.19.0
parse-type==0.6.0
pigar==2.0.2
Pillow==9.3.0
pluggy==1.0.0
protobuf==4.21.10
py==1.11.0
pycparser==2.21
pygit2==1.11.1
Pygments==2.13.0
pygount==1.4.0
pymsteams==0.2.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
python-dateutil==2.8.2
PyYAML==6.0
Random-Word==1.0.11
requests==2.28.1
retry==0.9.2
retrying==1.3.4
scheduler==0.8.2
selenium==4.7.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sure==2.0.0
svgwrite==1.4.3
text-unidecode==1.3
toml==0.10.2
traitlets==5.6.0
Tree==0.2.4
treelib==1.6.1
trio==0.22.0
trio-websocket==0.9.2
typeguard==2.13.3
urllib3==1.26.12
wsproto==1.2.0
